+++ Innowise.Music/Configuration/ApiSettings.cs
@@ -6,5 +6,9 @@ public class ApiSettings
     
     public string BaseUrl { get; set; } = string.Empty;
     public string AndroidBaseUrl { get; set; } = string.Empty;
+    
+    public string StreamBaseUrl { get; set; } = string.Empty;
+    public string AndroidStreamBaseUrl { get; set; } = string.Empty;
+    
     public int SearchPageSize { get; set; } = 8;
 }