+++ Innowise.Music/Services/RecommendationService.cs
@@ -54,8 +54,8 @@ public class RecommendationService : IRecommendationService
 
             // Use HTTP for stream URLs (MediaElement can't handle self-signed HTTPS certs)
             var streamBaseUrl = DeviceInfo.Platform == DevicePlatform.Android
-                ? "http://10.0.2.2:5236"
-                : "http://localhost:5236";
+                ? _apiSettings.AndroidStreamBaseUrl
+                : _apiSettings.StreamBaseUrl;
 
             var tracks = response?.Tracks?.Select(t => new Track
             {